requirements.txt
-e .
pytest==6.1.1
numpy==1.19.1
scipy==1.5.2
pandas==1.0.3
ipykernel==5.2.1
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
Jinja2==2.10.1
jupyter==1.0.0
jupyter-client==6.1.5
jupyter-console==6.1.0
jupyter-core==4.6.3
notebook==6.0.3
matplotlib==3.2.1
seaborn==0.10.1
tensorboard==2.2.1
mypy==0.770
pylint==2.6.0
simpleitk==1.2.4
torch-summary==1.2.0
torchio==0.17.46
pexpect==4.8.0
python-telegram-bot==12.8
opencv-python==4.4.0.44